+++ test/misskey_federation_test.dart
@@ -1,0 +1,6 @@
+import 'package:misskey_dart/misskey_dart.dart';
+import 'package:test/test.dart';
+
+import 'util/misskey_dart_test_util.dart';
+
+void main() async {}
